HyperAttentionDTI/requirements.txt
pytorch>=1.2
numpy
sklearn
tqdm
tensorboardX
prefetch_generator